% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hash.R
\name{ascii_val}
\alias{ascii_val}
\title{Calculate the ascii value number of a character}
\usage{
ascii_val(char)
}
\arguments{
\item{char}{character}
}
\value{
ascii value integer
}
\description{
Calculate the ascii value number of a character
}

