% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hash.R
\name{base16_to_dec}
\alias{base16_to_dec}
\title{Converts a base 16 string to a base 10 number.
Because I couldn't get base R functions to work for big hex numbers.}
\usage{
base16_to_dec(str_16)
}
\arguments{
\item{str_16}{base 16 number as a string.}
}
\value{
base 10 integer.
}
\description{
Converts a base 16 string to a base 10 number.
Because I couldn't get base R functions to work for big hex numbers.
}

