% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hash.R
\name{dec_to_base16}
\alias{dec_to_base16}
\title{Converts a base 10 number to base 16 number.
Because I couldn't get R's as.hexmode() to work for big integers.}
\usage{
dec_to_base16(dec)
}
\arguments{
\item{dec}{base 10 integer}
}
\value{
base 16 number as a string
}
\description{
Converts a base 10 number to base 16 number.
Because I couldn't get R's as.hexmode() to work for big integers.
}

