% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hash.R
\name{decode}
\alias{decode}
\title{Decodes a hashid into the original integer or integer vector}
\usage{
decode(hash_str, settings)
}
\arguments{
\item{hash_str}{hashid string to decode into integer or integer vector}

\item{settings}{Settings list generated by hashid_settings}
}
\value{
integer or integer vector
}
\description{
Decodes a hashid into the original integer or integer vector
}

