% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hash.R
\name{decode_hex}
\alias{decode_hex}
\title{Decodes a hashid into the original hexidecimal number}
\usage{
decode_hex(hashid, settings)
}
\arguments{
\item{hashid}{hashid to decode}

\item{settings}{Settings list generated by hashid_settings}
}
\value{
hexidecimal number as a string
}
\description{
Decodes a hashid into the original hexidecimal number
}

