% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hash.R
\name{encode}
\alias{encode}
\title{Encodes an integer or integer vector into a hashid string.
All numbers must be non-negative integers.}
\usage{
encode(int, settings)
}
\arguments{
\item{int}{Integer or integer vector to encode}

\item{settings}{Settings list generated by hashid_settings}
}
\value{
hashid string
}
\description{
Encodes an integer or integer vector into a hashid string.
All numbers must be non-negative integers.
}

