% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hash.R
\name{encode_hex}
\alias{encode_hex}
\title{Encodes a hexademical number into a hashid}
\usage{
encode_hex(hex_str, settings)
}
\arguments{
\item{hex_str}{Hexadecimal number as string}

\item{settings}{Settings list generated by hashid_settings}
}
\value{
hashid string
}
\description{
Encodes a hexademical number into a hashid
}

