% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hash.R
\name{enforce_min_length}
\alias{enforce_min_length}
\title{Enforces hashid minimum length by padding the hashid with additional characters.}
\usage{
enforce_min_length(encoded, min_length, alphabet, guards, values_hash)
}
\arguments{
\item{encoded}{encoded hashid}

\item{min_length}{minimum length required for hashid}

\item{alphabet}{set of letters used to generate hashid}

\item{guards}{set of guards used to generate hashid}

\item{values_hash}{value hashed used to select guard characters}
}
\value{
hashid with padded characters to insure minimum length
}
\description{
Enforces hashid minimum length by padding the hashid with additional characters.
}

