% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/defaults.R
\docType{data}
\name{hashid_defaults}
\alias{DEFAULT_ALPHABET}
\alias{DEFAULT_SEPS}
\alias{RATIO_GUARDS}
\alias{RATIO_SEPARATORS}
\alias{hashid_defaults}
\title{Default Values for hashid settings}
\format{\preformatted{ chr "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890"
}}
\source{
http://www.hashids.org
}
\usage{
DEFAULT_ALPHABET

DEFAULT_SEPS

RATIO_SEPARATORS

RATIO_GUARDS
}
\description{
Default alphabet, separators, and
   ratio of character separators and guards for hashid
}
\keyword{datasets}

