% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hash.R
\name{hashid_settings}
\alias{hashid_settings}
\title{A function to create a hashid settings list.}
\usage{
hashid_settings(salt, min_length = 0, alphabet = DEFAULT_ALPHABET,
  sep = DEFAULT_SEPS)
}
\arguments{
\item{salt}{An additional string to make hashids more unique.}

\item{min_length}{Minimum length for hashid.}

\item{alphabet}{String of characters for hashid.}

\item{sep}{String of characters to use as separators.}
}
\value{
A list of parameters used in encoding and decoding.
}
\description{
A function to create a hashid settings list.
}

