% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hash.R
\name{shuffle}
\alias{shuffle}
\title{Permutes the characters in a string based on an inputted salt string.}
\usage{
shuffle(string, salt)
}
\arguments{
\item{string}{String to be permuted}

\item{salt}{cryptograph salt string that is used to permute strings}
}
\value{
shuffled string
}
\description{
Permutes the characters in a string based on an inputted salt string.
}

