% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hash.R
\name{split}
\alias{split}
\title{Splits a string based on a set of splitting characters}
\usage{
split(string, splitters)
}
\arguments{
\item{string}{String to split}

\item{splitters}{set of splitting characters as a string}
}
\value{
split vector of characters
}
\description{
Splits a string based on a set of splitting characters
}

